Deprecation Logger: Improve how anonymous functions are logged, when applicable.
It would happen that an anonymous function had no relation to a file, but instead showed up as part of the stack trace and was never caught. This change adds a second pass over, in case no entries are found that matches plugins, themes, or WordPress core, that will return the anonymous function location, if available.